pyproject.toml
[tool.poetry]
name = "talentleads"
version = "0.1.0"
description = "Find the Best Talent for Your Startup"
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.7"
python-webpack-boilerplate = "^1.0.0"
django-widget-tweaks = "^1.4.12"
psycopg2 = "^2.9.5"
django-model-utils = "^4.3.1"
croniter = "^1.3.8"
django-autoslug = "^1.9.8"
markdown = "^3.4.1"
gunicorn = "^20.1.0"
pillow = "^12.0.0"
django-extensions = "^3.2.1"
django-filter = "^22.1"
django-anymail = {extras = ["mailgun"], version = "^9.1"}
pygments = "^2.14.0"
whitenoise = "^6.4.0"
django-taggit = "^3.1.0"
django-modelcluster = "^6.0"
libgravatar = "^1.0.4"
django-storages = "^1.13.2"
boto3 = "^1.26.137"
redis = {extras = ["hiredis"], version = "^4.6.0"}
openai = "^1.6.1"
sentry-sdk = "^2.42.0"
structlog-sentry = "^2.2.1"
django-structlog = "^9.1.1"
structlog = "^25.4.0"
django-q2 = "^1.8.0"
django-allauth = "^65.12.1"
django-environ = "^0.12.0"
dj-stripe = "^2.10.3"
pgvector = "^0.4.1"
pydantic-ai = "^1.1.0"
httpx = "^0.28.1"
logfire = "^4.13.2"
ipython = "^9.6.0"
stripe = "^13.0.1"
django-ninja = "^1.5.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
djlint = "^1.36.4"
ruff = "^0.14.1"
ty = "^0.0.1a23"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.djlint]
profile="django"
ignore = "H031,H006,H023,H011"

[tool.ruff]
line-length = 120
